Refactor for opt in usage and simplify API
