+++ packages/ra-core/src/core/CoreAdminRoutes.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect, Children, ComponentType } from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { WithPermissions, useCheckAuth } from '../auth';
-import { ApplicationUpdateMode, useTimeout } from '../util';
+import { useTimeout } from '../util';
 import { useScrollToTop, useCreatePath } from '../routing';
 import {
     AdminChildren,
@@ -35,7 +35,6 @@ export const CoreAdminRoutes = (props: CoreAdminRoutesProps) => {
         requireAuth,
         ready: Ready,
         title,
-        updateMode,
     } = props;
 
     const [canRender, setCanRender] = useState(!requireAuth);
@@ -78,12 +77,7 @@ export const CoreAdminRoutes = (props: CoreAdminRoutesProps) => {
                 path="/*"
                 element={
                     <div>
-                        <Layout
-                            dashboard={dashboard}
-                            menu={menu}
-                            title={title}
-                            updateMode={updateMode}
-                        >
+                        <Layout dashboard={dashboard} menu={menu} title={title}>
                             <Routes>
                                 {customRoutesWithLayout}
                                 {Children.map(resources, resource => (
@@ -132,7 +126,6 @@ export interface CoreAdminRoutesProps extends CoreLayoutProps {
     loading: LoadingComponent;
     requireAuth?: boolean;
     ready?: ComponentType;
-    updateMode?: ApplicationUpdateMode;
 }
 
 const defaultAuthParams = { params: { route: 'dashboard' } };
