+++ packages/ra-core/src/core/CoreAdminUI.tsx
@@ -3,7 +3,7 @@ import { ComponentType, useEffect, isValidElement, createElement } from 'react';
 import { Routes, Route } from 'react-router-dom';
 
 import { CoreAdminRoutes } from './CoreAdminRoutes';
-import { ApplicationUpdateMode, Ready } from '../util';
+import { Ready } from '../util';
 import {
     TitleComponent,
     LoginComponent,
@@ -35,7 +35,6 @@ export interface CoreAdminUIProps {
     requireAuth?: boolean;
     ready?: ComponentType;
     title?: TitleComponent;
-    updateMode?: ApplicationUpdateMode;
 }
 
 export const CoreAdminUI = (props: CoreAdminUIProps) => {
@@ -52,7 +51,6 @@ export const CoreAdminUI = (props: CoreAdminUIProps) => {
         ready = Ready,
         title = 'React Admin',
         requireAuth = false,
-        updateMode,
     } = props;
 
     useEffect(() => {
@@ -94,7 +92,6 @@ export const CoreAdminUI = (props: CoreAdminUIProps) => {
                         requireAuth={requireAuth}
                         ready={ready}
                         title={title}
-                        updateMode={updateMode}
                     >
                         {children}
                     </CoreAdminRoutes>
