+++ packages/ra-core/src/types.ts
@@ -1,7 +1,6 @@
 import { ComponentType, ReactElement, ReactNode } from 'react';
 import { WithPermissionsChildrenParams } from './auth/WithPermissions';
 import { AuthActionType } from './auth/types';
-import { ApplicationUpdateMode } from './util';
 
 /**
  * data types
@@ -323,7 +322,6 @@ export interface CoreLayoutProps {
         hasDashboard?: boolean;
     }>;
     title?: TitleComponent;
-    updateMode?: ApplicationUpdateMode;
 }
 
 export type LayoutComponent = ComponentType<CoreLayoutProps>;
