+++ packages/ra-core/src/util/index.ts
@@ -31,5 +31,4 @@ export * from './shallowEqual';
 export * from './LabelPrefixContext';
 export * from './LabelPrefixContextProvider';
 export * from './useLabelPrefix';
-export * from './CheckForApplicationUpdate';
 export * from './useCheckForApplicationUpdate';
