+++ packages/ra-core/src/util/useCheckForApplicationUpdate.ts
@@ -10,35 +10,30 @@ import { useEvent } from './useEvent';
  * @param {string} options.onNewVersionAvailable The function to call when a new version of the application is available.
  * @param {string} options.url Optional. The URL to download to check for code update. Defaults to the HTML page.
  * @param {number} options.checkInterval Optional. The interval in milliseconds between two checks. Defaults to 3600000 (1 hour).
+ * @param {boolean} options.disabled Optional. Whether the check should be disabled. Defaults to false.
  */
 export const useCheckForApplicationUpdate = (
     options: UseCheckForApplicationUpdateOptions
 ) => {
-    const shouldDisable =
-        process.env.NODE_ENV !== 'production' ||
-        // Some automation tools may trigger security warnings when the app tries to download itself
-        // so we disable the feature when running on localhost to still allow production builds to be tested locally
-        window.location.hostname === 'localhost';
-
     const {
         url = window.location.href,
         checkInterval = ONE_HOUR,
         onNewVersionAvailable,
-        updateMode = shouldDisable ? 'disabled' : 'manual',
+        disabled,
     } = options;
     const currentHash = useRef<string>();
     const onCodeHasChanged = useEvent(onNewVersionAvailable);
 
     useEffect(() => {
-        if (updateMode === 'disabled') return;
+        if (disabled) return;
 
         getHashForUrl(url).then(hash => {
             currentHash.current = hash;
         });
-    }, [updateMode, url]);
+    }, [disabled, url]);
 
     useEffect(() => {
-        if (updateMode === 'disabled') return;
+        if (disabled) return;
 
         const interval = setInterval(() => {
             getHashForUrl(url).then(hash => {
@@ -48,7 +43,7 @@ export const useCheckForApplicationUpdate = (
             });
         }, checkInterval);
         return () => clearInterval(interval);
-    }, [checkInterval, onCodeHasChanged, updateMode, url]);
+    }, [checkInterval, onCodeHasChanged, disabled, url]);
 };
 
 const getHashForUrl = async (url: string) => {
@@ -73,7 +68,5 @@ export interface UseCheckForApplicationUpdateOptions {
     onNewVersionAvailable: () => void;
     checkInterval?: number;
     url?: string;
-    updateMode?: ApplicationUpdateMode;
+    disabled?: boolean;
 }
-
-export type ApplicationUpdateMode = 'disabled' | 'immediate' | 'manual';
